fix: added declarative siblingMatchers
