+++ resources/kindhandlers/handlers/cert-manager.io/Certificate.json
@@ -18,6 +18,10 @@
       "source": {
         "pathParts": [
           "spec", "issuerRef", "name"
+        ],
+        "matchers": [
+          "kindMatcher",
+          "groupMatcher"
         ]
       },
       "type": "name",
